refactor: runtime env variables
