+++ assets/entrepot/api/alerts.ts
@@ -1,10 +1,10 @@
+import { annexesUrl, communityId } from "@/env";
 import { IApiAlert } from "../../@types/alert";
 import { jsonFetch } from "../../modules/jsonFetch";
 
-const communityId = (document.getElementById("root") as HTMLDivElement)?.dataset?.["configCommunityId"];
 const annexePath = "/public/alerts.json";
 const fileName = annexePath.substring(annexePath.lastIndexOf("/") + 1);
-const baseUrl = "https://data.geopf.fr/annexes/cartes.gouv.fr-config";
+const baseUrl = `${annexesUrl}/cartes.gouv.fr-config`;
 
 const get = (otherOptions: RequestInit = {}) => {
     return jsonFetch<IApiAlert[]>(`${baseUrl}${annexePath}`, {
