+++ assets/pages/Home.tsx
@@ -11,8 +11,9 @@ import mapSvgUrl from "@codegouvfr/react-dsfr/dsfr/artwork/pictograms/map/map.sv
 import systemSvgUrl from "@codegouvfr/react-dsfr/dsfr/artwork/pictograms/system/system.svg?no-inline";
 import { useEffect } from "react";
 
+import { catalogueUrl } from "@/env";
 import SymfonyRouting from "@/modules/Routing";
-import { catalogueUrl, routes, useRoute } from "@/router/router";
+import { routes, useRoute } from "@/router/router";
 import { useAuthStore } from "@/stores/AuthStore";
 import homeImgUrl from "@/img/home/home.png";
 import Main from "@/components/Layout/Main";
